refactor: LottoNumber 를 VO 로 이동
