+++ src/main/java/lotto/LottoApplication.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 import lotto.domain.Lotto;
 import lotto.domain.LottoGroup;
-import lotto.domain.LottoNumber;
 import lotto.domain.LottoResult;
 import lotto.domain.WinningLotto;
+import lotto.domain.vo.LottoNumber;
 import lotto.domain.vo.Money;
 import lotto.domain.vo.Quantity;
 import lotto.input.LottoInput;
