+++ src/main/java/lotto/domain/WinningLotto.java
@@ -1,6 +1,7 @@
 package lotto.domain;
 
 import java.util.List;
+import lotto.domain.vo.LottoNumber;
 
 public class WinningLotto {
